Pull 1.7 release notes images into the docs tree.
The images used for the 1.7 release notes were all hosted on reviewboard.org.
This was both annoying in terms of things being self-contained (we'd have to go
upload things separately from writing the docs), and the format of them was
pretty bad when building the docs to a format other than HTML.

I've downloaded all of the images and referenced them locally instead of
remotely. In addition, I've added some conditionals to include the full images
when the target isn't HTML and fixed the latex target.

Testing done:
* Built the HTML docs and verified that the thumbnails showed up, and that
  links to the full versions went to the correct place.
* Built the latex docs, compiled to PDF, and verified that the full screenshots
  were included rather than the thumbnails.

Reviewed at https://reviews.reviewboard.org/r/4646/