Expand our REST API test suite, and simplify the codebase.
This enhances the API test suite to automate a bunch of the common tests
for resources. The standard GET, POST, PUT, DELETE tests that test basic
behavior, with local sites, and without access to local sites have all
been generalized through mixins. Tests for resources that exist as
direct or indirect children of ReviewRequestResource also have
additional tests that test for access when review requests are made
private.

Much of the test suite code has been removed in favor of the new
mixed in tests. An additional 20+ seconds of tests have been added.