Fix several access and payload issues with the REST API.
We had a handful of functions that weren't properly checking for valid
access. This only affected installations using local sites and private
review requests, and required knowledge of table IDs, but it was still
possible to get access to some information. This change ensures we have
the proper decorators on all functions (and removes some redundant
decorators), and that access permissions are checked in all cases.

There were also some issues in returning 404 when resources couldn't be
accessed, and a case where the wrong payload was returned. These have
been fixed as well.

This change does affect extensions that provide API resources. They will
need to provide a has_access_permisisons method that returns True if the
user should be able to access the resource. That's no longer going to
allow access by default, as it's now opt-in.