Tweak the heck out of the repository search box.
The search box had some drawing problems, especially when scrollbars were
visible. I hadn't noticed this during the initial development because OSX.

This change tweaks the heck out of the javascript and CSS to try to make things
draw more reliably. The "Repositories" header is now hidden when the search box
is visible, to make it less likely that things will overlap. I've also tried to
tweak the positioning code to make it work better when scrollbars are present.

This is still not 100% perfect, but I'm about ready to hunt down the box model
spec authors and stab them in the face, so I'm going to leave it here for now.

Testing done:
Played around with this in Chrome, Safari, and Firefox, using a handful of zoom
levels. Turned on scrollbars always and saw that things were improved (though
not perfect).

Reviewed at https://reviews.reviewboard.org/r/4700/