Add extensibility to the comment dialog.
The comment dialog is now extensible through a CommentDialogHook. Such
hooks provide a view that will essentially layer onto the CommentDialog
and can manipulate or listen to it or its associated CommentEditor.

Some of the internals of CommentDialog have been made public for use by
extensions.

Right now, hooks can manipulate the buttons and trigger operations such
as saving. Later, we'll probably want to allow it to add additional
fields.