Protect against comment objects without reviews when loading a review UI.
Somehow, in my database I ended up with comments that were not
associated with a Review. While this is unlikely in production, it's
good to check for this, as it'll otherwise result in a confusing
JavaScript syntax error when rendering the template.

Now, when loading comments, we catch exceptions when attempting to get
the review, and ignore the comment.