Add support for JavaScript-side extensions to Review Board.
A Review Board extension can now define a JavaScript counterpart that
can plug into hooks in JavaScript, in order to augment functionality.

Previously, an extension would have to define a TemplateHook with a
script to load, which would then have to monkey-patch its way into
things. We're moving beyond that now to provide a solid way of hooking
in.