Add a hook for displaying extra comment information for reviews.
This introduces a new hook, CommentDetailDisplayHook, which can be used
to display additional data in reviews and in e-mails (both HTML and
plain text). An extension can use this to show information such as, say,
the severity of a comment.

This does not provide the ability to display in the review dialog, as
that's generated in JavaScript entirely. That will be added in a future
change.