Add a hook for working with comments in the review dialog.
This new hook, ReviewDialogCommentHook, allows an extension to display
data or provide UI for a comment in the review dialog. It takes a view
type that will later be instantiated and passed the comment as a model.
The view can do whatever it needs to with the model, including providing
UI to set new data on it.