Don't trigger revisionSelected from the revision selector when there's no change.
Dragging a handle on the revision selector would cause a reload, even if it was
dropped back on the place that it started. This change guards it, and fixes
some logic to copy the values array instead of referencing it.

Testing done:
Checked that changing the revision selector would trigger a load, except when
the handle was dropped back on its previous location.

Reviewed at https://reviews.reviewboard.org/r/4746/