Support the new static bundles for extensions.
This updates to support the new static bundles for extensions. The
appropriate template tags have been placed in the base template, so that
the default bundles will be loaded.

There's also a new reviewboard.extensions.packaging module, which
contains a setup function that an extension's setup.py can use in place
of the standard setuptools one. This will set some required options for
the package, and set things up to install static bundles.

The documentation and extension generator have been updated.