Support static media bundles in review UIs.
ReviewUI now has css_bundle_names and js_bundle_names variables, which
specify static media bundles that will be loaded in when displaying the
review UI. This makes it easier to write a Review UI without having to
use a custom template to include the required files.

Extensions can use Extension.get_bundle_id to pass in bundle IDs for a
review UI, but it must do it in the ReviewUI's __init__ function, or
patch the class in the Extension's __init__, as the extension IDs (and
therefore bundle IDs) are computed once an extension is enabled and
constructed, not when defining the classes.