Fix the fill-database command.
Changes in the diff code as well as django-1.5 caused fill-database to break.
This change fixes it, switching to use the UploadDiffForm from the reviews
module, and calling is_valid() to make sure that the form cleans the input
data.

Testing done:
Ran the command and saw that it actually ran and populated the database
correctly. Ran the unit tests, they all pass.

Reviewed at https://reviews.reviewboard.org/r/4692/