Make the Markdown styles consistent with the source text.
The Markdown styles no longer jump nearly as much when going from
editing to viewing. The text aligns nearly everywhere except when there
are non-visible Markdown formatting characters, like * or `.

The CSS we use for the diff viewer syntax highlighting is now being used
for source code in Markdown, which replaces prettify.css.

Also, in general, styles are just a bit cleaner and more tightened.

Due to some strange bugs in Firefox with its textarea, where things are
off by 1px in places, some Firefox-specific CSS rules had to be added to
compensate.