Reorganize the list of resources in the webapi docs.
The webapi docs had a flat, alphabetical list of resources, which wasn't
that easy to navigate. This was especially true since we had things like
"file attachment comments," "review file attachment comments," "review
reply file attachment comments," etc.

Now, all resources in this list are grouped by sections, making it much
easier to locate a particular category of resources and to see how they
relate.

It also deletes file-attachment-comment, which isn't even really a
resource and should never have been there (since the list for that
resource forwards on to the an instance within the parent review in the
tree).