Document extension API resources.
Extensions could add their own resources to the WebAPI, but this was never
documented. I've added a sample of this that looks a lot like Review Bot's
review resource, but simplified quite a bit to eliminate a lot of the actual
guts, leaving only a shell of a resource.

Testing done: Generated HTML docs.
Reviewed at https://reviews.reviewboard.org/r/4818/