Fix a breakage with interdiffs.
The new interdiff code had a few occasions where it'd miss chunks. It
didn't actually miss them, it just miscategorized as equals.

The reason it did this was that it was attempting to figure out the next
range to consider after figuring out if a chunk was valid, and was only
considering the very next range. What it didn't consider was that there
may be a handful of ranges it should be skipping over that are equal
between the interdiffs, and that it may need to figure this out before
determining if a chunk is valid

This fixes the broken interdiff noticed on
https://reviews.reviewboard.org/r/4257/.

A unit test was added for this case.