Fix overzealous escaping of periods when escaping for Markdown.
The period is listed as an escapable character, but it's really only
important after numbers at the beginning of a line. The regexes for
escaping now account for this, and don't escape periods elsewhere.

It's very likely we'll find we want to be even more thorough on some
other escaping, such as parens, but this is the most noticeable one.