Document Markdown support.
This change adds some relatively thorough documentation for using markdown.
The bulk of this is a page called "Using Markdown" which gives a quick overview
of the basics of markdown syntax for the things most people will care about. I
haven't made this an exhaustive document, since the pages on daring fireball
and github serve that purpose. Instead, this should hopefully be a nice quick
reference for the basics of markdown syntax as interpreted by Review Board.

This is linked to from all the various places where markdown can be used.

While I was in here, I updated everything to capitalize "Markdown", including
in the MD review UI docs, and capitalized the terms in the glossary (which
looks much better to me).

Testing done:
* Generated HTML docs and read through everything.
* Checked cross-linking.

Reviewed at https://reviews.reviewboard.org/r/4834/