Show a spinner before rendering markdown text or diff index.
When displaying a review request, there's a brief moment of time before
Markdown gets rendered or before the diff table is loaded where things
look wrong. For the Markdown fields (namely the Description and Testing
Done fields), the raw Markdown is shown first, and then switches to the
rendered text, which looks bad. The diff index shows a big line before
displaying its contents.

These now are marked as loading at first, and we'll show spinners. Once
the page has successfully loaded, the spinners go away and the rendered
content appears.

We also try to render the text for these fields a bit sooner than we did
before, hopefully reducing the load time.