Fix the grip image on the comment dialog.
When we added the jquery-ui stylesheet, and upgraded jquery-ui, it broke
our custom grip image on the comment dialog. That image remained, but an
arrow overlayed it.

This brings it back, and in a cleaner way. We don't add it through
JavaScript anymore. Instead, we just set the CSS rule for that icon.

It now looks as it did in earlier versions.