Improve the style of the move indicator flags.
The move indicator flags were very noisy things. There was a flag for
every moved line, which made it harder to see that a range was moved as
a whole, if two ranges were right up against each other.

This changes the design to only have a flag for the first line in a
range, and not subsequent lines. Those are indicated by a tail on the
flag, or a flag pole perhaps.

This reduces the visual noise considerably, at the expense of only
allowing the first row to be clickable to jump lines. That's a necessity
due to the limitations in how we can represent this style in CSS (as we
cannot properly represent each row of that flag pole as an element with
the proper height in a TD without javascript).