Lighten the deleted lines color in the diff viewer.
The deleted lines color in the diff viewer was still a lot darker than
the insert and replace lines, and the contrast with the move flags
weren't great. This lightens it a bit more, making it appear a lot
closer to the other colors, and making it easier to read.