Change default timezone to UTC
Commit 107ce3e3721c5517f345595a5740f2108e40e015 changed the pytz requirement to
be >= 2012h because versions earlier than that do not support the US/Pacific
designation of the timezone, which Review Board was using as a default. This
makes it difficult or impossible to install Review Board on Linux distributions
with an older version of pytz.

This patch changes the default timezone to UTC, which should be available in
all versions of pytz since time out of memory. With this, we can drop the pytz
strict version requirement as well, trusting that the platform version will
support UTC.