Improve diff pagination a little bit.
This doesn't change pagination in a wild way, it just fixes it to not be super
weird. During the diff index changes, the pagination links at the top of the
diffs disappeared, leaving it so the only one was at the bottom. This was a
rude surprise when I slogged through a big review only to discover that there
were two or three more pages! I've restored the pagination links at the top,
which means there's one set of links after the diff index and one at the very
bottom of the page.

I also changed it so switching pages will operate without a reload. This makes
it very smooth to navigate around.

Testing done:
Switched around to a bunch of different revisions and interdiff revisions with
multiple pages. Made sure that switching pages always went to the right place,
and that switching to new revisions always loaded the first page of the
new revision.

Reviewed at https://reviews.reviewboard.org/r/4856/