Switch the position of the moved-line flags.
One thing about the new moved-lines detection that bugged me a lot is that the
flags were on the wrong side of the diff columns ("moved from" was on the left
hand side with an arrow pointing to the right, and "moved to" was on the right
with an arrow pointing to the left). This really grated against my intution
that the left hand side was "from" and the right hand side was "to".

I've therefore switched which side each of the flags appears on. This has the
side effect of increasing the amount of vertical space used, because most of
the time we now have to show a flagpole. That said, I think it's a highly
valuable trade-off.

Testing done:
Created some review request with moved sections and saw that things made way
more sense than before.

Reviewed at https://reviews.reviewboard.org/r/4863/