Pass on all exceptions when creating new review requests with diffs.
In most cases, errors parsing the diff file or communicating with the
repository got turned into form errors which would be shown to the user.
Unfortunately, DiffParserError wasn't accounted for, since it inherits directly
from Exception. I've removed the long list of specific exceptions and just made
it catch Exception directly, since we handle everything lower down in the form.