Clean up the diff loading style.
The diff loading style was a bit broken-looking. The "Revision" and
"New Change" headers jumped on load, and the spinner had no padding
around it.

It's impossible to properly align the revision headers without knowing
some of the visible line numbers and content and placing it in the
table, and any best-guesses are just guesses. The next best thing is to
just not show revision information until the file has loaded, which is
what we do now. It's a bit cleaner-looking, with just a filename and a
spinner (which now has some padding around it as well).