Move review request draft banner logic into new classes.
This moves all the logic for dealing with draft banners into a set of
new classes in ReviewRequestEditorView. These will allow
ReviewRequestEditorView to dynamically create these banners, instead of
requiring a page reload.

Right now, a page reload is still used, so as not to make this change
any larger than it already is. Dynamic updating will come in a future
change.