Turn on the X-Frame-Options header for clickjacking prevention.
Django provides middleware to turn on the X-Frame-Options header, which
hints to the browsers (those that support it) to prevent loading if
inside a frame (in our case, in any frame not from the same origin).
This will be the default for new Django projects in 1.6, but worth
turning it on here.

This is not a security vulnerability. Rather, this is just extra
hardening, for the browsers that choose to support this header.