Revamp the review request layout.
We've had the same look to the review request box since the dawn of
time. With 2.0, we're changing that to take up less vertical space,
to bring a little more consistency with some of the fields, and to
highlight the main content of the review request.

The main information (Summary, Description and Testing Done) are now
shown one after another, with the same font style used for the labels.

On the right-hand side is the list of fields we previously had between
Summary and Description. They're a single column now. This takes the
space that was often just left unused within the Description/Testing
Done fields (since those fields were often far wider than the contents).

Simplifying this into one column will also make it easier to lay out
fields provided by hooks.

The Screenshots, File Attachments and Issue Summary are displayed after
all the fields, as before. Their labels match the style of the Summary,
Description and Testing Done labels.

The Issue Summary table has been cleaned up a bit to have the filters
and the counters on the same row, reducing the amount of vertical space.